Removed LazyReference now that Delay is better, stronger, faster.
