+++ lib/concurrent.rb
@@ -19,7 +19,6 @@ require 'concurrent/delay'
 require 'concurrent/exchanger'
 require 'concurrent/future'
 require 'concurrent/ivar'
-require 'concurrent/lazy_reference'
 require 'concurrent/lazy_register'
 require 'concurrent/mvar'
 require 'concurrent/promise'
