+++ lib/concurrent/utility/processor_count.rb
@@ -1,12 +1,12 @@
 require 'rbconfig'
-require 'concurrent/lazy_reference'
+require 'concurrent/delay'
 
 module Concurrent
 
   class ProcessorCounter
     def initialize
-      @processor_count          = LazyReference.new { compute_processor_count }
-      @physical_processor_count = LazyReference.new { compute_physical_processor_count }
+      @processor_count          = Delay.new { compute_processor_count }
+      @physical_processor_count = Delay.new { compute_physical_processor_count }
     end
 
     # Number of processors seen by the OS and used for process scheduling. For
